refactor: modify product API structures
